Split out compare script
